Move _split_address_and_target_path unit tests to create_test.py
Problem: The test_split_address_* tests are pure unit tests for an isolated
function, but they were placed in test_create.py (the integration test file)
instead of create_test.py (the unit test file).
Fix: Moved the 7 tests and the import to create_test.py, matching the
convention used by similar tests like test_parse_branch_flag_*.
