+++ libs/mngr/imbue/mngr/cli/test_create.py
@@ -15,7 +15,6 @@ from imbue.mngr.api.agent_addr import parse_agent_address
 from imbue.mngr.cli.create import _create_agent
 from imbue.mngr.cli.create import _resolve_transfer_mode
 from imbue.mngr.cli.create import _setup_create
-from imbue.mngr.cli.create import _split_address_and_target_path
 from imbue.mngr.cli.create import create
 from imbue.mngr.config.data_types import CreateCliOptions
 from imbue.mngr.config.data_types import MngrContext
@@ -1163,60 +1162,6 @@ def test_transfer_git_worktree_rejected_for_non_git(
 
     assert result.exit_code != 0
     assert "git repository" in result.output.lower()
-
-
-# =============================================================================
-# Tests for _split_address_and_target_path
-# =============================================================================
-
-
-def test_split_address_no_colon() -> None:
-    """Plain address with no colon returns no target path."""
-    addr, path = _split_address_and_target_path("foo")
-    assert addr == "foo"
-    assert path is None
-
-
-def test_split_address_empty_string() -> None:
-    """Empty string returns empty address and no target path."""
-    addr, path = _split_address_and_target_path("")
-    assert addr == ""
-    assert path is None
-
-
-def test_split_address_with_absolute_path() -> None:
-    """Address with :PATH returns address and path."""
-    addr, path = _split_address_and_target_path("foo:/tmp/work")
-    assert addr == "foo"
-    assert path == Path("/tmp/work")
-
-
-def test_split_address_with_relative_path() -> None:
-    """Address with relative :PATH works."""
-    addr, path = _split_address_and_target_path(":./rel/path")
-    assert addr == ""
-    assert path == Path("./rel/path")
-
-
-def test_split_address_full_address_with_path() -> None:
-    """Full agent address with :PATH suffix."""
-    addr, path = _split_address_and_target_path("foo@host.modal:/root/work")
-    assert addr == "foo@host.modal"
-    assert path == Path("/root/work")
-
-
-def test_split_address_path_only() -> None:
-    """Colon-prefixed path with no address."""
-    addr, path = _split_address_and_target_path(":/tmp/work")
-    assert addr == ""
-    assert path == Path("/tmp/work")
-
-
-def test_split_address_trailing_colon() -> None:
-    """Trailing colon with no path returns no target path."""
-    addr, path = _split_address_and_target_path("foo:")
-    assert addr == "foo"
-    assert path is None
 
 
 def test_transfer_none_with_different_target_path_rejected(
